requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.2
mysql_connector_repackaged==0.3.1
PyYAML==5.4.1
PyYAML==6.0.1
Requests==2.31.0
